Remove redundant Storage Alter in workspace, as now Storage Usage fulfills this role
