--- frontend/src/components/common/StorageAlert.tsx
@@ -1,336 +1,0 @@
-import React, { useEffect, useState } from "react"
-
-import { useSnackbar } from "notistack"
-
-import {
-  Close as CloseIcon,
-  Refresh as RefreshIcon,
-  Storage as StorageIcon,
-  Warning as WarningIcon,
-} from "@mui/icons-material"
-import {
-  Alert,
-  AlertTitle,
-  Box,
-  Button,
-  CircularProgress,
-  Collapse,
-  IconButton,
-  LinearProgress,
-  Typography,
-} from "@mui/material"
-
-import {
-  getMyStorageAlertApi,
-  getMyStorageUsageApi,
-  refreshStorageUsageApi,
-  StorageAlert as StorageAlertType,
-  StorageUsage,
-} from "api/storage/StorageAlerts"
-import { SubscriptionAlertThresholds } from "const/Subscription"
-
-interface StorageAlertProps {
-  showUsageDetails?: boolean
-  compact?: boolean
-  onClose?: () => void
-}
-
-const StorageAlert: React.FC<StorageAlertProps> = ({
-  showUsageDetails = false,
-  compact = false,
-  onClose,
-}) => {
-  const { enqueueSnackbar } = useSnackbar()
-  const [alert, setAlert] = useState<StorageAlertType | null>(null)
-  const [usage, setUsage] = useState<StorageUsage | null>(null)
-  const [loading, setLoading] = useState(true)
-  const [refreshing, setRefreshing] = useState(false)
-  const [dismissed, setDismissed] = useState(false)
-  const [error, setError] = useState<string | null>(null)
-
-  const fetchStorageAlert = async () => {
-    try {
-      setLoading(true)
-      setError(null)
-
-      // Fetch alert and usage independently so one failure doesn't block the other
-      const [alertResult, usageResult] = await Promise.allSettled([
-        getMyStorageAlertApi(),
-        showUsageDetails ? getMyStorageUsageApi() : Promise.resolve(null),
-      ])
-
-      if (alertResult.status === "fulfilled" && alertResult.value) {
-        const alertResponse = alertResult.value
-        if (alertResponse.has_alert && alertResponse.alert) {
-          setAlert(alertResponse.alert)
-        } else {
-          setAlert(null)
-        }
-      }
-
-      if (usageResult.status === "fulfilled" && usageResult.value) {
-        setUsage(usageResult.value)
-      }
-
-      // If both failed, show error
-      if (
-        alertResult.status === "rejected" &&
-        usageResult.status === "rejected"
-      ) {
-        const msg =
-          alertResult.reason instanceof Error
-            ? alertResult.reason.message
-            : "Unknown error"
-        // eslint-disable-next-line no-console
-        console.error("Failed to fetch storage data:", msg)
-        setError("Unable to load storage information")
-      }
-    } catch (err) {
-      // eslint-disable-next-line no-console
-      console.error("Failed to fetch storage alert:", err)
-      setError("Unable to load storage information")
-    } finally {
-      setLoading(false)
-    }
-  }
-
-  const handleRefresh = async () => {
-    try {
-      setRefreshing(true)
-      await refreshStorageUsageApi()
-      await fetchStorageAlert()
-      enqueueSnackbar("Storage usage refreshed", { variant: "success" })
-    } catch (error) {
-      // eslint-disable-next-line no-console
-      console.error("Failed to refresh storage:", error)
-      enqueueSnackbar("Failed to refresh storage usage", { variant: "error" })
-    } finally {
-      setRefreshing(false)
-    }
-  }
-
-  const handleDismiss = () => {
-    setDismissed(true)
-    onClose?.()
-  }
-
-  useEffect(() => {
-    fetchStorageAlert()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [showUsageDetails])
-
-  if (loading) {
-    return compact ? null : (
-      <Box display="flex" alignItems="center" gap={1} p={1}>
-        <CircularProgress size={16} />
-        <Typography variant="caption">Checking storage...</Typography>
-      </Box>
-    )
-  }
-
-  if (dismissed || (!alert && !showUsageDetails)) {
-    return null
-  }
-
-  if (error && !alert && !usage) {
-    return (
-      <Alert
-        severity="warning"
-        sx={{ mb: 2 }}
-        action={
-          <Button size="small" onClick={fetchStorageAlert}>
-            Retry
-          </Button>
-        }
-      >
-        {error}
-      </Alert>
-    )
-  }
-
-  const getSeverity = (alertLevel: string) => {
-    switch (alertLevel) {
-      case "danger":
-        return "error"
-      case "critical":
-        return "error"
-      case "warning":
-        return "warning"
-      default:
-        return "info"
-    }
-  }
-
-  const getProgressColor = (percentage: number) => {
-    if (percentage >= SubscriptionAlertThresholds.CRITICAL) return "error"
-    if (percentage >= SubscriptionAlertThresholds.WARNING) return "warning"
-    return "primary"
-  }
-
-  if (compact && alert) {
-    return (
-      <Collapse in={!dismissed}>
-        <Alert
-          severity={getSeverity(alert.alert_level)}
-          action={
-            <Box display="flex" alignItems="center" gap={0.5}>
-              <IconButton
-                size="small"
-                onClick={handleRefresh}
-                disabled={refreshing}
-              >
-                {refreshing ? <CircularProgress size={16} /> : <RefreshIcon />}
-              </IconButton>
-              <IconButton size="small" onClick={handleDismiss}>
-                <CloseIcon />
-              </IconButton>
-            </Box>
-          }
-        >
-          <AlertTitle>
-            <Box display="flex" alignItems="center" gap={1}>
-              <StorageIcon fontSize="small" />
-              Storage{" "}
-              {alert.alert_level === "danger" ? "Quota Exceeded" : "Alert"}
-            </Box>
-          </AlertTitle>
-          {alert.message}
-        </Alert>
-      </Collapse>
-    )
-  }
-
-  return (
-    <Box>
-      {alert && (
-        <Collapse in={!dismissed}>
-          <Alert
-            severity={getSeverity(alert.alert_level)}
-            action={
-              <IconButton size="small" onClick={handleDismiss}>
-                <CloseIcon />
-              </IconButton>
-            }
-            sx={{ mb: 2 }}
-          >
-            <AlertTitle>
-              <Box display="flex" alignItems="center" gap={1}>
-                <StorageIcon />
-                Storage{" "}
-                {alert.alert_level === "danger" ? "Quota Exceeded" : "Alert"}
-              </Box>
-            </AlertTitle>
-            <Typography variant="body2" sx={{ mb: 1 }}>
-              {alert.message}
-            </Typography>
-
-            <Box display="flex" alignItems="center" gap={2} mt={1}>
-              <LinearProgress
-                variant="determinate"
-                value={Math.min(alert.storage_usage_percent, 100)}
-                color={getProgressColor(alert.storage_usage_percent)}
-                sx={{ flexGrow: 1, height: 8, borderRadius: 4 }}
-              />
-              <Typography variant="caption" fontWeight="bold">
-                {alert.storage_usage_percent.toFixed(1)}%
-              </Typography>
-            </Box>
-
-            <Box
-              display="flex"
-              justifyContent="space-between"
-              alignItems="center"
-              mt={1}
-            >
-              <Typography variant="caption" color="text.secondary">
-                {alert.storage_usage_bytes > 0
-                  ? `${(alert.storage_usage_bytes / (1024 * 1024 * 1024)).toFixed(2)} GB used of ${(alert.storage_quota_bytes / (1024 * 1024 * 1024)).toFixed(2)} GB`
-                  : "Calculating usage..."}
-              </Typography>
-              <Button
-                size="small"
-                startIcon={
-                  refreshing ? <CircularProgress size={16} /> : <RefreshIcon />
-                }
-                onClick={handleRefresh}
-                disabled={refreshing}
-              >
-                Refresh
-              </Button>
-            </Box>
-          </Alert>
-        </Collapse>
-      )}
-
-      {showUsageDetails && usage && (
-        <Box
-          sx={{
-            border: 1,
-            borderColor: "divider",
-            borderRadius: 1,
-            p: 2,
-            mt: alert ? 0 : 2,
-          }}
-        >
-          <Box
-            display="flex"
-            alignItems="center"
-            justifyContent="space-between"
-            mb={2}
-          >
-            <Typography variant="body1" color="text.secondary">
-              Storage Usage
-            </Typography>
-          </Box>
-
-          {usage.storage_quota_bytes ? (
-            <>
-              <Box display="flex" alignItems="center" gap={2} mb={2}>
-                <LinearProgress
-                  variant="determinate"
-                  value={Math.min(usage.storage_usage_percent || 0, 100)}
-                  color={getProgressColor(usage.storage_usage_percent || 0)}
-                  sx={{ flexGrow: 1, height: 12, borderRadius: 6 }}
-                />
-                <Typography variant="body2" fontWeight="bold" minWidth="60px">
-                  {usage.storage_usage_percent?.toFixed(1) || "0.0"}%
-                </Typography>
-              </Box>
-
-              <Box display="flex" justifyContent="space-between" mb={1}>
-                <Typography variant="body2" color="text.secondary">
-                  Used: {usage.storage_usage_formatted}
-                </Typography>
-                <Typography variant="body2" color="text.secondary">
-                  Total: {usage.storage_quota_formatted}
-                </Typography>
-              </Box>
-
-              {usage.alert_level && (
-                <Box display="flex" alignItems="center" gap={1} mt={2}>
-                  <WarningIcon
-                    color={usage.alert_level === "danger" ? "error" : "warning"}
-                    fontSize="small"
-                  />
-                  <Typography
-                    variant="caption"
-                    color={usage.alert_level === "danger" ? "error" : "warning"}
-                  >
-                    Storage usage is{" "}
-                    {usage.alert_level === "danger" ? "over quota" : "high"}
-                  </Typography>
-                </Box>
-              )}
-            </>
-          ) : (
-            <Typography variant="body2" color="text.secondary">
-              {usage.storage_usage_formatted} used (no quota limit set)
-            </Typography>
-          )}
-        </Box>
-      )}
-    </Box>
-  )
-}
-
-export default StorageAlert
