+++ frontend/src/pages/Workspace/index.tsx
@@ -39,7 +39,7 @@ import { refreshAllWorkspacesStorageApi } from "api/workspace"
 import DeleteConfirmModal from "components/common/DeleteConfirmModal"
 import Loading from "components/common/Loading"
 import PaginationCustom from "components/common/PaginationCustom"
-import StorageAlert from "components/common/StorageAlert"
+import StorageUsage from "components/common/StorageUsage"
 import PopupShare from "components/Workspace/PopupShare"
 import { selectCurrentUser } from "store/slice/User/UserSelector"
 import { resetVisualizeLayout } from "store/slice/VisualizeItem/VisualizeItemSlice"
@@ -642,11 +642,7 @@ const Workspaces = () => {
           New
         </Button>
       </Box>
-      <StorageAlert
-        key={storageRefreshKey}
-        showUsageDetails={true}
-        compact={false}
-      />
+      <StorageUsage key={storageRefreshKey} />
       {user ? (
         <Box
           sx={{
